Crate gnuplot

A simple gnuplot controller.

Example

use gnuplot::{Figure, Caption, Color};

let x = [0u32, 1, 2];
let y = [3u32, 4, 5];
let mut fg = Figure::new();
fg.axes2d()
.lines(&x, &y, &[Caption("A line"), Color("black")]);
fg.show();

Re-exports

pub use self::Coordinate::*;

pub use self::AlignType::*;

pub use self::ArrowheadType::*;

pub use self::AutoOption::*;

pub use self::BorderLocation2D::*;

pub use self::ContourStyle::*;

pub use self::DashType::*;

pub use self::FillRegionType::*;

pub use self::LabelOption::*;

pub use self::LegendOption::*;

pub use self::MarginSide::*;

pub use self::PaletteType::*;

pub use self::PlotOption::*;

pub use self::Tick::*;

pub use self::TickOption::*;

Structs

Axes2D
2D axes that is used for drawing 2D plots

Axes3D
3D axes that is used for drawing 3D plots

Figure
A figure that may contain multiple axes

Enums

AlignType
An enumeration of possible text and label alignments

ArrowheadType
An enumeration of possible arrow head styles

AutoOption
An enumeration of something that can either be fixed (e.g. the maximum of X values), or automatically determined

BorderLocation2D
Plot border locations

ContourStyle
Specifies how the contours are drawn

Coordinate
Specifies how to interpret the coordinate passed to a plotting command

DashType
An enumeration of possible dash styles

FillRegionType
An enumeration of possible fill regions

LabelOption
An enumeration of label options that control label attributes

LegendOption
Legend options

MarginSide
Plot margins

PaletteType
Specifies what sort of palette to use

PlotOption
An enumeration of plot options you can supply to plotting commands, governing things like line width, color and others

Tick
Specifies a type of axis tick

TickOption
An enumeration of axis tick options

Constants

COLOR
Default Gnuplot palette

GRAY
A gray palette

HELIX
A nice default for a cube helix

HOT
A black body palette

RAINBOW
Classic rainbow palette with terrible perceptual properties

Traits

AxesCommon

DataType