Enum gnuplot::coordinates::Coordinate

pub enum Coordinate {
    Graph(f64),
    Axis(f64),
}

Specifies how to interpret the coordinate passed to a plotting command

Variants

Graph

Coordinates are done relative to a graph (i.e. an axis set). (0, 0) is the bottom left corner and (1, 1) is the top right corner. You'd use this to place labels and other objects so that they remain in the same place relative to the graph no matter what you have plotted.

Axis

Coordinates match those on the axes. You'd use this to place labels and other objects relative to regions of interest in the graph (e.g. labeling the peak of a function)