Enum gnuplot::LabelOption

pub enum LabelOption<'l> {
    TextOffset(f64, f64),
    Font(&'l str, f64),
    TextColor(&'l str),
    Rotate(f64),
    TextAlign(AlignType),
    MarkerSymbol(char),
    MarkerColor(&'l str),
    MarkerSize(f64),
}

An enumeration of label options that control label attributes

Variants

TextOffset

Sets the offset of the label in characters

Font

Sets the font of the label. The string specifies the font type (e.g. "Arial") and the number specifies the size (the units are terminal dependent, but are often points)

TextColor

Sets the color of the label text. The passed string can be a color name (e.g. "black" works), or an HTML color specifier (e.g. "#FFFFFF" is white)

Rotate

Rotates the label by a certain number of degrees

TextAlign

Sets the horizontal alignment of the label text (default is left alignment). See AlignType.

MarkerSymbol

Sets a marker for the label. By default no marker is drawn. The valid characters are as follows:

. - dot

+ - plus

x - cross

* - star

s - empty square

S - filled square

o - empty circle

O - filled circle

t - empty triangle

T - filled triangle

d - empty del (upside down triangle)

D - filled del (upside down triangle)

r - empty rhombus

R - filled rhombus

MarkerColor

Sets the color of the marker. The passed string can be a color name (e.g. "black" works), or an HTML color specifier (e.g. "#FFFFFF" is white)

MarkerSize

Sets the size of the marker. The size acts as a multiplier, with 1.0 being the default.