Enum gnuplot::LegendOption

pub enum LegendOption<'l> {
    Reverse,
    Invert,
    Horizontal,
    Placement(AlignType, AlignType),
    Title(&'l str),
    MaxRows(u32),
    MaxCols(u32),
}

Legend options

Variants

Reverse

Puts curve samples to the left of the label

Invert

Displays legend entries in opposite order

Horizontal

Makes the legend horizontal (default is vertical)

Placement

Specifies the location of the legend. The first argument specifies the horizontal placement with respect to its position, and the second argument specifies the vertical placement

Title

Title of the legend

MaxRows

Specifies the maximum number of rows, when the legend is vertical

MaxCols

Specifies the maximum number of columns, when the legend is horizontal