Enum gnuplot::PlotOption

pub enum PlotOption<T> {
    PointSymbol(char),
    PointSize(f64),
    Caption(T),
    LineWidth(f64),
    Color(T),
    BorderColor(T),
    LineStyle(DashType),
    FillAlpha(f64),
    FillRegion(FillRegionType),
    ArrowType(ArrowheadType),
    ArrowSize(f64),
    WhiskerBars(f64),
}

An enumeration of plot options you can supply to plotting commands, governing things like line width, color and others

Variants
PointSymbol(char)

Sets the symbol used for points. The valid characters are as follows:

. - dot

+ - plus

x - cross

* - star

s - empty square

S - filled square

o - empty circle

O - filled circle

t - empty triangle

T - filled triangle

d - empty del (upside down triangle)

D - filled del (upside down triangle)

r - empty rhombus

R - filled rhombus

PointSize(f64)

Sets the size of the points. The size acts as a multiplier, with 1.0 being the default.

Caption(T)

Sets the caption of the plot element. Set to empty to hide it from the legend.

LineWidth(f64)

Sets the width of lines.

Color(T)

Sets the color of the plot element. The passed string can be a color name (e.g. "black" works), or an HTML color specifier (e.g. "#FFFFFF" is white). This specifies the fill color of a filled plot.

BorderColor(T)

Sets the color of the border of a filled plot (if it has one). The passed string can be a color name (e.g. "black" works), or an HTML color specifier (e.g. "#FFFFFF" is white).

LineStyle(DashType)

Sets the style of the line. Note that not all gnuplot terminals support dashed lines. See DashType for the available styles.

FillAlpha(f64)

Sets the transparency of a filled plot. 0.0 - fully transparent, 1.0 - fully opaque

FillRegion(FillRegionType)

Sets the fill region. See FillRegion for the available regions.

ArrowType(ArrowheadType)

Sets what an arrowhead looks like

ArrowSize(f64)

Sets the size of the arrowhead. This is specified in the units of graph (i.e. 1.0 would make the arrow as big as the graph).

WhiskerBars(f64)

Width of the whisker bars (as a fraction of the box width) for box and whisker plots.