Enum gnuplot​::options​::TickOption​

pub enum TickOption {
    OnAxis(bool),
    Mirror(bool),
    Inward(bool),
    MinorScale(f64),
    MajorScale(f64),
}

An enumeration of axis tick options

Variants

OnAxis

Specifies whether the ticks are drawn at the borders of the plot, or on the axis

Mirror

If the axes are drawn on the border, this specifies whether to draw the ticks on the opposite border as well

Inward

If the axes are drawn on the border, this specifies whether to draw the ticks pointing inward or outward

MinorScale

Sets the scale of the minor ticks

MajorScale

Sets the scale of the major ticks