Enum gnuplot::PaletteType

pub enum PaletteType {
    Gray(f32),
    Formula(i32, i32, i32),
    CubeHelix(f32, f32, f32, f32),
}

Specifies what sort of palette to use

Variants

Gray

Use a gray palette with a specified gamma

Formula

Use a palette with that uses a predefined formula for each color component. Each formula is identified by an integer between [-36, 36]. See gnuplot documentation, or use the pre-defined constants.

CubeHelix

Use a cube helix palette, with a certain start (in radians), cycles, saturation and gamma.