Enum gnuplot::PlotOption

pub enum PlotOption<'l> {
    PointSymbol(char),
    PointSize(f64),
    Caption(&'l str),
    LineWidth(f64),
    Color(&'l str),
    BorderColor(&'l str),
    LineStyle(DashType),
    FillAlpha(f64),
    FillRegion(FillRegionType),
    ArrowType(ArrowheadType),
    ArrowSize(f64),
}

An enumeration of plot options you can supply to plotting commands, governing things like line width, color and others

Variants

PointSymbol

Sets the symbol used for points. The valid characters are as follows:

. - dot

+ - plus

x - cross

* - star

s - empty square

S - filled square

o - empty circle

O - filled circle

t - empty triangle

T - filled triangle

d - empty del (upside down triangle)

D - filled del (upside down triangle)

r - empty rhombus

R - filled rhombus

PointSize

Sets the size of the points. The size acts as a multiplier, with 1.0 being the default.

Caption

Sets the caption of the plot element. Set to empty to hide it from the legend.

LineWidth

Sets the width of lines.

Color

Sets the color of the plot element. The passed string can be a color name (e.g. "black" works), or an HTML color specifier (e.g. "#FFFFFF" is white). This specifies the fill color of a filled plot.

BorderColor

Sets the color of the border of a filled plot (if it has one). The passed string can be a color name (e.g. "black" works), or an HTML color specifier (e.g. "#FFFFFF" is white).

LineStyle

Sets the style of the line. Note that not all gnuplot terminals support dashed lines. See DashType for the available styles.

FillAlpha

Sets the transparency of a filled plot. 0.0 - fully transparent, 1.0 - fully opaque

FillRegion

Sets the fill region. See FillRegion for the available regions.

ArrowType

Sets what an arrowhead looks like

ArrowSize

Sets the size of the arrowhead. This is specified in the units of graph (i.e. 1.0 would make the arrow as big as the graph).