Struct gnuplot::Axes3D

pub struct Axes3D { /* fields omitted */ }

3D axes that is used for drawing 3D plots

Methods

impl Axes3D

pub fn surface<'l, T: DataType, X: IntoIterator<Item = T>>(
&'l mut self,
mat: X,
num_rows: usize,
num_cols: usize,
dimensions: Option<(f64, f64, f64, f64)>,
options: &[PlotOption<&str>]
) -> &'l mut Self

Draws a 3D surface from a rectangular array of data by connecting the individual datapoints with polygons.

#Arguments:

mat - Row-major 2D array signifying the Z coordinate of the datapoints. The X and Y coordinates of the datapoints are determined automatically, and optionally scaled using the dimensions argument.

num_rows - Number of rows in the data array

num_cols - Number of columns in the data array

dimensions - Optional X and Y coordinates of the first and last data points (with the rest of the coordinates spaced evenly between). By default this will be (0, 0) and (num_rows - 1, num_cols - 1).

options - Array of PlotOption controlling the appearance of the surface. Relevant options are:

Caption - Specifies the caption for this dataset. Use an empty string to hide it (default).

pub fn points<'l, Tx: DataType, X: IntoIterator<Item = Tx>, Ty: DataType, Y: IntoIterator<Item = Ty>, Tz: DataType, Z: IntoIterator<Item = Tz>>(
&'l mut self,
x: X,
y: Y,
z: Z,
options: &[PlotOption<&str>]
) -> &'l mut Self

Plot a 3D scatter-plot with a point standing in for each data point

Arguments

x - x values

y - y values

z - z values

options - Array of PlotOption<&str> controlling the appearance of the plot element. The relevant options are:

Caption - Specifies the caption for this dataset. Use an empty string to hide it (default).

PointSymbol - Sets symbol for each point

PointSize - Sets the size of each point

Color - Sets the color

pub fn lines<'l, Tx: DataType, X: IntoIterator<Item = Tx>, Ty: DataType, Y: IntoIterator<Item = Ty>, Tz: DataType, Z: IntoIterator<Item = Tz>>(
&'l mut self,
x: X,
y: Y,
z: Z,
options: &[PlotOption<&str>]
) -> &'l mut Self

Plot a 3D scatter-plot with lines connecting each data point

Arguments

x - x values

y - y values

z - z values

options - Array of PlotOption<&str> controlling the appearance of the plot element. The relevant options are:

Caption - Specifies the caption for this dataset. Use an empty string to hide it (default).

PointSymbol - Sets symbol for each point

PointSize - Sets the size of each point

Color - Sets the color

pub fn lines_points<'l, Tx: DataType, X: IntoIterator<Item = Tx>, Ty: DataType, Y: IntoIterator<Item = Ty>, Tz: DataType, Z: IntoIterator<Item = Tz>>(
&'l mut self,
x: X,
y: Y,
z: Z,
options: &[PlotOption<&str>]
) -> &'l mut Self

A combination of lines and points methods (drawn in that order).

Arguments

x - x values

y - y values

z - z values

options - Array of PlotOption<&str> controlling the appearance of the plot element

pub fn set_view<'l>(&'l mut self, pitch: f64, yaw: f64) -> &'l mut Self

Sets the 3D view.

#Arguments:

pitch - Pitch, in degrees. Value of 0 is looking straight down on the XY plane, Z pointing out of the screen.

yaw - Yaw, in degrees. Value of 0 is looking at the XZ plane, Y point into the screen.

pub fn set_view_map<'l>(&'l mut self) -> &'l mut Self

Sets the view to be a map. Useful for images and contour plots.

pub fn set_z_label<'l>(
&'l mut self,
text: &str,
options: &[LabelOption<&str>]
) -> &'l mut Self

Set the label for the Z axis

Arguments

text - Text of the label. Pass an empty string to hide the label

options - Array of LabelOption controlling the appearance of the label. Relevant options are:

Offset - Specifies the offset of the label

Font - Specifies the font of the label

TextColor - Specifies the color of the label

Rotate - Specifies the rotation of the label

Align - Specifies how to align the label

pub fn set_z_ticks<'l>(
&'l mut self,
tick_placement: Option<(AutoOption<f64>, u32)>,
tick_options: &[TickOption<&str>],
label_options: &[LabelOption<&str>]
) -> &'l mut Self

Like set_x_ticks but for the Z axis.

pub fn set_z_ticks_custom<'l, T: DataType, TL: IntoIterator<Item = Tick<T>>>(
&'l mut self,
ticks: TL,
tick_options: &[TickOption<&str>],
label_options: &[LabelOption<&str>]
) -> &'l mut Self

Like set_x_ticks_custom but for the the Y axis.

pub fn set_z_range<'l>(
&'l mut self,
min: AutoOption<f64>,
max: AutoOption<f64>
) -> &'l mut Self

Set the range of values for the Z axis

Arguments

min - Minimum Z value

max - Maximum Z value

pub fn set_z_reverse<'l>(&'l mut self, reverse: bool) -> &'l mut Self

Sets z axis to reverse.

pub fn set_z_log<'l>(&'l mut self, base: Option<f64>) -> &'l mut Self

Sets the Z axis be logarithmic. Note that the range must be non-negative for this to be valid.

Arguments

base - If Some, then specifies base of the logarithm, if None makes the axis not be logarithmic

pub fn set_z_grid<'l>(&'l mut self, show: bool) -> &'l mut Self

Shows the grid on the Z axis.

Arguments

show - Whether to show the grid.

pub fn show_contours<'l>(
&'l mut self,
base: bool,
surface: bool,
style: ContourStyle,
label: AutoOption<&str>,
levels: AutoOption<u32>
) -> &'l mut Self

Show contours (lines of equal Z value) at automatically determined levels.

Arguments

base - Show contours on the base of the plot (XY plane)

surface - Show the contours on the surface itself

style - Style of the contours

label - Auto sets the label automatically and enables the legend, Fix() allows you specify a format string (using C style formatting), otherwise an empty string disables the legend and labels.

levels - Auto picks some default number of levels, otherwise you can pass a set nominal number instead. The number is nominal as contours are placed at nice values of Z, and thus there may be fewer of them than this number.

pub fn show_contours_custom<'l, T: DataType, TC: IntoIterator<Item = T>>(
&'l mut self,
base: bool,
surface: bool,
style: ContourStyle,
label: AutoOption<&str>,
levels: TC
) -> &'l mut Self

Show contours (lines of equal Z value) at specific levels.

Arguments

base - Show contours on the base of the plot (XY plane)

surface - Show the contours on the surface itself

style - Style of the contours

label - Auto sets the label automatically and enables the legend, Fix() allows you specify a format string (using C style formatting), otherwise an empty string disables the legend and labels.

levels - A set of levels.